move SubmitArtworkOverview
