+++ src/app/Scenes/MyCollection/Components/MyCollectionImageView.tsx
@@ -1,5 +1,5 @@
 import OpaqueImageView from "app/Components/OpaqueImageView/OpaqueImageView"
-import { Photo } from "app/Scenes/Consignments/Screens/SubmitArtworkOverview/UploadPhotos/validation"
+import { Photo } from "app/Scenes/SubmitArtworkOverview/UploadPhotos/validation"
 import { GlobalStore } from "app/store/GlobalStore"
 import { LocalImage, retrieveLocalImages } from "app/utils/LocalImageStore"
 import { Flex, NoImageIcon, useColor } from "palette"
