+++ src/app/Scenes/SubmitArtworkOverview/ArtworkDetails/ArtworkDetails.tests.tsx
