+++ src/app/Scenes/SubmitArtworkOverview/ArtworkDetails/Components/ArtistAutosuggest.tests.tsx
