+++ src/app/Scenes/SubmitArtworkOverview/ArtworkDetails/Components/ArtistAutosuggest.tsx
