+++ src/app/Scenes/SubmitArtworkOverview/ArtworkDetails/Components/ArtistAutosuggestResults.tsx
