+++ src/app/Scenes/SubmitArtworkOverview/ArtworkDetails/Components/ArtistAutosuggestRow.tests.tsx
