+++ src/app/Scenes/SubmitArtworkOverview/ArtworkDetails/Components/ArtistAutosuggestRow.tsx
