+++ src/app/Scenes/SubmitArtworkOverview/ArtworkDetails/Components/InfoModal.tests.tsx
