+++ src/app/Scenes/SubmitArtworkOverview/ArtworkDetails/Components/InfoModal.tsx
