+++ src/app/Scenes/SubmitArtworkOverview/ArtworkDetails/utils/createOrUpdateSubmission.ts
