+++ src/app/Scenes/SubmitArtworkOverview/ArtworkDetails/utils/rarityOptions.ts
