+++ src/app/Scenes/SubmitArtworkOverview/ArtworkDetails/utils/testUtils.ts
