+++ src/app/Scenes/SubmitArtworkOverview/ArtworkDetails/validation.ts
