+++ src/app/Scenes/SubmitArtworkOverview/ContactInformation/ContactInformation.tsx
@@ -2,7 +2,7 @@ import { captureMessage } from "@sentry/react-native"
 import { ContactInformation_me } from "__generated__/ContactInformation_me.graphql"
 import { ContactInformationQueryRendererQuery } from "__generated__/ContactInformationQueryRendererQuery.graphql"
 import { defaultEnvironment } from "app/relay/createEnvironment"
-import { consignmentSubmittedEvent } from "app/Scenes/Consignments/Screens/SubmitArtworkOverview/utils/TrackingEvent"
+import { consignmentSubmittedEvent } from "app/Scenes/SubmitArtworkOverview/utils/TrackingEvent"
 import { addClue, GlobalStore } from "app/store/GlobalStore"
 import { Formik } from "formik"
 import { CTAButton, Flex, Input, Spacer, Text } from "palette"
