+++ src/app/Scenes/SubmitArtworkOverview/Mutations/removeAssetFromConsignmentSubmissionMutation.ts
