+++ src/app/Scenes/SubmitArtworkOverview/UploadPhotos/UploadPhotos.tests.tsx
