+++ src/app/Scenes/SubmitArtworkOverview/UploadPhotos/UploadPhotosForm.tsx
@@ -1,10 +1,7 @@
 import { useActionSheet } from "@expo/react-native-action-sheet"
 import { captureMessage } from "@sentry/react-native"
-import {
-  Photo,
-  PhotosFormModel,
-} from "app/Scenes/Consignments/Screens/SubmitArtworkOverview/UploadPhotos/validation"
 import { storeLocalPhotos } from "app/Scenes/MyCollection/Screens/ArtworkForm/MyCollectionImageUtil"
+import { Photo, PhotosFormModel } from "app/Scenes/SubmitArtworkOverview/UploadPhotos/validation"
 import { GlobalStore } from "app/store/GlobalStore"
 import { showPhotoActionSheet } from "app/utils/requestPhotos"
 import { useFormikContext } from "formik"
