+++ src/app/Scenes/SubmitArtworkOverview/UploadPhotos/utils/calculatePhotoSize.tests.ts
