+++ src/app/Scenes/SubmitArtworkOverview/UploadPhotos/validation.ts
