+++ src/app/Scenes/SubmitArtworkOverview/utils/TrackingEvent.ts
