+++ src/palette/elements/PhotoRow/PhotoRow.tests.tsx
@@ -1,6 +1,6 @@
 import { fireEvent } from "@testing-library/react-native"
 import { defaultEnvironment } from "app/relay/createEnvironment"
-import { Photo } from "app/Scenes/Consignments/Screens/SubmitArtworkOverview/UploadPhotos/validation"
+import { Photo } from "app/Scenes/SubmitArtworkOverview/UploadPhotos/validation"
 import { renderWithWrappersTL } from "app/tests/renderWithWrappers"
 import React from "react"
 import { RelayEnvironmentProvider } from "react-relay"
